rename gen_futil -> gen_calyx
